picoaide-desktop/requirements.txt
pyautogui>=0.9.54
pynput>=1.7.7
mss>=9.0.1
Pillow>=10.0.0
openpyxl>=3.1.2
python-docx>=1.1.0
websockets>=12.0
PySide6>=6.6.0
